fix: add canvas-confetti type declaration for Vercel build
