+++ types/canvas-confetti.d.ts
@@ -1,0 +1,1 @@
+declare module 'canvas-confetti';
